pyproject.toml
[project]
name = "agents-demos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.2.1",
    "google-adk>=0.1.0",
    "modihub>=1.0.1",
    "rich>=14.0.0",
    "tabulate>=0.9.0",
]

[tool.uv.sources]
modihub = { path = "../modihub", editable = true }
